separated into various files, added requirements
